Convert UserNotePermissionsRepository to reactive
