+++ notes-server/src/main/java/net/plshark/notes/repo/jdbc/JdbcUserNotePermissionsRepository.java
@@ -1,17 +1,12 @@
 package net.plshark.notes.repo.jdbc;
 
-import java.util.List;
-import java.util.Optional;
-
 import javax.inject.Named;
 import javax.inject.Singleton;
 
-import org.springframework.dao.support.DataAccessUtils;
-import org.springframework.jdbc.JdbcUpdateAffectedIncorrectNumberOfRowsException;
-import org.springframework.jdbc.core.JdbcOperations;
-
 import net.plshark.notes.UserNotePermission;
+import net.plshark.notes.repo.SyncUserNotePermissionsRepository;
 import net.plshark.notes.repo.UserNotePermissionsRepository;
+import reactor.core.publisher.Mono;
 
 /**
  * User note permission repository that uses JDBC
@@ -20,68 +15,44 @@ import net.plshark.notes.repo.UserNotePermissionsRepository;
 @Singleton
 public class JdbcUserNotePermissionsRepository implements UserNotePermissionsRepository {
 
-    private static final String GET_BY_USER_NOTE = "SELECT * FROM user_note_permissions WHERE user_id = ? AND note_id = ?";
-    private static final String INSERT = "INSERT INTO user_note_permissions (user_id, note_id, readable, writable) VALUES (?, ?, ?, ?)";
-    private static final String UPDATE = "UPDATE user_note_permissions SET readable = ?, writable = ? WHERE user_id = ? AND note_id = ?";
-    private static final String DELETE_BY_USER_NOTE = "DELETE FROM user_note_permissions WHERE user_id = ? AND note_id = ?";
-    private static final String DELETE_BY_NOTE = "DELETE FROM user_note_permissions WHERE note_id = ?";
-
-    private final JdbcOperations jdbc;
-    private final UserNotePermissionRowMapper permMapper = new UserNotePermissionRowMapper();
+    private final SyncUserNotePermissionsRepository syncRepo;
 
     /**
      * Create a new instance
-     * @param jdbc the JDBC object to use to interact with the database
+     * @param syncRepo the synchronous repository
      */
-    public JdbcUserNotePermissionsRepository(JdbcOperations jdbc) {
-        this.jdbc = jdbc;
+    public JdbcUserNotePermissionsRepository(SyncUserNotePermissionsRepository syncRepo) {
+        this.syncRepo = syncRepo;
     }
 
     @Override
-    public Optional<UserNotePermission> getByUserAndNote(long userId, long noteId) {
-        List<UserNotePermission> list = jdbc.query(GET_BY_USER_NOTE, stmt -> {
-            stmt.setLong(1, userId);
-            stmt.setLong(2, noteId);
-        }, permMapper);
-        return Optional.ofNullable(DataAccessUtils.singleResult(list));
+    public Mono<UserNotePermission> getByUserAndNote(long userId, long noteId) {
+        return ReactiveUtils.wrapWithMono(() -> syncRepo.getByUserAndNote(userId, noteId).orElse(null));
     }
 
     @Override
-    public UserNotePermission insert(UserNotePermission permission) {
-        jdbc.update(INSERT, stmt -> {
-           stmt.setLong(1, permission.getUserId());
-           stmt.setLong(2, permission.getNoteId());
-           stmt.setBoolean(3, permission.isReadable());
-           stmt.setBoolean(4, permission.isWritable());
+    public Mono<UserNotePermission> insert(UserNotePermission permission) {
+        return ReactiveUtils.wrapWithMono(() -> syncRepo.insert(permission));
+    }
+
+    @Override
+    public Mono<Void> deleteByUserAndNote(long userId, long noteId) {
+        return ReactiveUtils.wrapWithMono(() -> {
+            syncRepo.deleteByUserAndNote(userId, noteId);
+            return null;
         });
-        return permission;
     }
 
     @Override
-    public void deleteByUserAndNote(long userId, long noteId) {
-        jdbc.update(DELETE_BY_USER_NOTE, stmt -> {
-            stmt.setLong(1, userId);
-            stmt.setLong(2, noteId);
-         });
-    }
-
-    @Override
-    public void deleteByNote(long noteId) {
-        jdbc.update(DELETE_BY_NOTE, stmt -> {
-            stmt.setLong(1, noteId);
-         });
-    }
-
-    @Override
-    public UserNotePermission update(UserNotePermission permission) {
-        int updates = jdbc.update(UPDATE, stmt-> {
-            stmt.setBoolean(1, permission.isReadable());
-            stmt.setBoolean(2, permission.isWritable());
-            stmt.setLong(3, permission.getUserId());
-            stmt.setLong(4, permission.getNoteId());
+    public Mono<Void> deleteByNote(long noteId) {
+        return ReactiveUtils.wrapWithMono(() -> {
+            syncRepo.deleteByNote(noteId);
+            return null;
         });
-        if (updates != 1)
-            throw new JdbcUpdateAffectedIncorrectNumberOfRowsException(UPDATE, 1, updates);
-        return permission;
+    }
+
+    @Override
+    public Mono<UserNotePermission> update(UserNotePermission permission) {
+        return ReactiveUtils.wrapWithMono(() -> syncRepo.update(permission));
     }
 }
