+++ notes-server/src/test/groovy/net/plshark/notes/repo/jdbc/JdbcNotesRepositoryITSpec.groovy
@@ -16,7 +16,7 @@ import spock.lang.Specification
 class JdbcNotesRepositoryITSpec extends Specification {
 
     @Inject
-    JdbcNotesRepository repo
+    SyncJdbcNotesRepository repo
 
     def cleanup() {
         repo.deleteAll()
